session: fix data race in the TestUpgradeVersion83 and TestUpgradeVersion75
