+++ session/bootstrap_test.go
@@ -18,7 +18,6 @@ import (
 	"context"
 	"fmt"
 	"strconv"
-	"strings"
 	"testing"
 
 	"github.com/pingcap/tidb/config"
@@ -32,6 +31,11 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+// This test file have many problem.
+// 1. Please use testkit to create dom, session and store.
+// 2. Don't use createStoreAndBootstrap and BootstrapSession together. It will cause data race.
+// Please do not add any test here. You can add test case at the update_test.go. After All problem fixed,
+// We will overwrite this file by update_test.go.
 func TestBootstrap(t *testing.T) {
 	store, dom := createStoreAndBootstrap(t)
 	defer func() { require.NoError(t, store.Close()) }()
@@ -776,85 +780,6 @@ func TestUpgradeVersion74(t *testing.T) {
 			row := req.GetRow(0)
 			require.Equal(t, strconv.Itoa(ca.newValue), row.GetString(0))
 		}()
-	}
-}
-
-func TestUpgradeVersion75(t *testing.T) {
-	ctx := context.Background()
-
-	store, _ := createStoreAndBootstrap(t)
-	defer func() { require.NoError(t, store.Close()) }()
-
-	seV74 := createSessionAndSetID(t, store)
-	txn, err := store.Begin()
-	require.NoError(t, err)
-	m := meta.NewMeta(txn)
-	err = m.FinishBootstrap(int64(74))
-	require.NoError(t, err)
-	err = txn.Commit(context.Background())
-	require.NoError(t, err)
-	mustExec(t, seV74, "update mysql.tidb set variable_value='74' where variable_name='tidb_server_version'")
-	mustExec(t, seV74, "commit")
-	mustExec(t, seV74, "ALTER TABLE mysql.user DROP PRIMARY KEY")
-	mustExec(t, seV74, "ALTER TABLE mysql.user MODIFY COLUMN Host CHAR(64)")
-	mustExec(t, seV74, "ALTER TABLE mysql.user ADD PRIMARY KEY(Host, User)")
-	unsetStoreBootstrapped(store.UUID())
-	ver, err := getBootstrapVersion(seV74)
-	require.NoError(t, err)
-	require.Equal(t, int64(74), ver)
-	r := mustExec(t, seV74, `desc mysql.user`)
-	req := r.NewChunk(nil)
-	row := req.GetRow(0)
-	require.NoError(t, r.Next(ctx, req))
-	require.Equal(t, "host", strings.ToLower(row.GetString(0)))
-	require.Equal(t, "char(64)", strings.ToLower(row.GetString(1)))
-
-	domV75, err := BootstrapSession(store)
-	require.NoError(t, err)
-	defer domV75.Close()
-	seV75 := createSessionAndSetID(t, store)
-	ver, err = getBootstrapVersion(seV75)
-	require.NoError(t, err)
-	require.Equal(t, currentBootstrapVersion, ver)
-	r = mustExec(t, seV75, `desc mysql.user`)
-	req = r.NewChunk(nil)
-	row = req.GetRow(0)
-	require.NoError(t, r.Next(ctx, req))
-	require.Equal(t, "host", strings.ToLower(row.GetString(0)))
-	require.Equal(t, "char(255)", strings.ToLower(row.GetString(1)))
-}
-
-func TestUpgradeVersion83(t *testing.T) {
-	ctx := context.Background()
-	store, _ := createStoreAndBootstrap(t)
-	defer func() { require.NoError(t, store.Close()) }()
-
-	domV83, err := BootstrapSession(store)
-	require.NoError(t, err)
-	defer domV83.Close()
-	seV83 := createSessionAndSetID(t, store)
-	ver, err := getBootstrapVersion(seV83)
-	require.NoError(t, err)
-	require.Equal(t, currentBootstrapVersion, ver)
-
-	statsHistoryTblFields := []struct {
-		field string
-		tp    string
-	}{
-		{"table_id", "bigint(64)"},
-		{"stats_data", "longblob"},
-		{"seq_no", "bigint(64)"},
-		{"version", "bigint(64)"},
-		{"create_time", "datetime(6)"},
-	}
-	rStatsHistoryTbl := mustExec(t, seV83, `desc mysql.stats_history`)
-	req := rStatsHistoryTbl.NewChunk(nil)
-	require.NoError(t, rStatsHistoryTbl.Next(ctx, req))
-	require.Equal(t, 5, req.NumRows())
-	for i := 0; i < 5; i++ {
-		row := req.GetRow(i)
-		require.Equal(t, statsHistoryTblFields[i].field, strings.ToLower(row.GetString(0)))
-		require.Equal(t, statsHistoryTblFields[i].tp, strings.ToLower(row.GetString(1)))
 	}
 }
