+++ session/upgrade_test.go
@@ -1,0 +1,105 @@
+// Copyright 2022 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package session_test
+
+import (
+	"context"
+	"strings"
+	"testing"
+
+	"github.com/pingcap/tidb/meta"
+	"github.com/pingcap/tidb/session"
+	"github.com/pingcap/tidb/testkit"
+	"github.com/stretchr/testify/require"
+)
+
+func TestUpgradeVersion83(t *testing.T) {
+	ctx := context.Background()
+	store, _, clean := testkit.CreateMockStoreAndDomain(t)
+	defer clean()
+
+	tk := testkit.NewTestKit(t, store)
+	ver, err := session.GetBootstrapVersion(tk.Session())
+	require.NoError(t, err)
+	require.Equal(t, session.CurrentBootstrapVersion, ver)
+
+	statsHistoryTblFields := []struct {
+		field string
+		tp    string
+	}{
+		{"table_id", "bigint(64)"},
+		{"stats_data", "longblob"},
+		{"seq_no", "bigint(64)"},
+		{"version", "bigint(64)"},
+		{"create_time", "datetime(6)"},
+	}
+	rStatsHistoryTbl, err := tk.Exec(`desc mysql.stats_history`)
+	require.NoError(t, err)
+	req := rStatsHistoryTbl.NewChunk(nil)
+	require.NoError(t, rStatsHistoryTbl.Next(ctx, req))
+	require.Equal(t, 5, req.NumRows())
+	for i := 0; i < 5; i++ {
+		row := req.GetRow(i)
+		require.Equal(t, statsHistoryTblFields[i].field, strings.ToLower(row.GetString(0)))
+		require.Equal(t, statsHistoryTblFields[i].tp, strings.ToLower(row.GetString(1)))
+	}
+}
+
+func TestUpgradeVersion75(t *testing.T) {
+	ctx := context.Background()
+
+	store, _, clean := testkit.CreateMockStoreAndDomain(t)
+	defer clean()
+
+	tk := testkit.NewTestKit(t, store)
+
+	seV74 := tk.Session()
+	txn, err := store.Begin()
+	require.NoError(t, err)
+	m := meta.NewMeta(txn)
+	err = m.FinishBootstrap(int64(74))
+	require.NoError(t, err)
+	err = txn.Commit(context.Background())
+	require.NoError(t, err)
+
+	tk.MustExec("update mysql.tidb set variable_value='74' where variable_name='tidb_server_version'")
+	tk.MustExec("commit")
+	tk.MustExec("ALTER TABLE mysql.user DROP PRIMARY KEY")
+	tk.MustExec("ALTER TABLE mysql.user MODIFY COLUMN Host CHAR(64)")
+	tk.MustExec("ALTER TABLE mysql.user ADD PRIMARY KEY(Host, User)")
+	session.UnsetStoreBootstrapped(store.UUID())
+	ver, err := session.GetBootstrapVersion(seV74)
+	require.NoError(t, err)
+	require.Equal(t, int64(74), ver)
+	r, err := tk.Exec(`desc mysql.user`)
+	require.NoError(t, err)
+	req := r.NewChunk(nil)
+	row := req.GetRow(0)
+	require.NoError(t, r.Next(ctx, req))
+	require.Equal(t, "host", strings.ToLower(row.GetString(0)))
+	require.Equal(t, "char(64)", strings.ToLower(row.GetString(1)))
+
+	tk2 := testkit.NewTestKit(t, store)
+	ver, err = session.GetBootstrapVersion(tk2.Session())
+	require.NoError(t, err)
+	require.Equal(t, session.CurrentBootstrapVersion, ver)
+	r, err = tk2.Exec(`desc mysql.user`)
+	require.NoError(t, err)
+	req = r.NewChunk(nil)
+	row = req.GetRow(0)
+	require.NoError(t, r.Next(ctx, req))
+	require.Equal(t, "host", strings.ToLower(row.GetString(0)))
+	require.Equal(t, "char(255)", strings.ToLower(row.GetString(1)))
+}
